requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiodns==3.2.0
    # via pydis-core
aiohttp==3.9.4
    # via
    #   -r requirements/requirements.in
    #   discord-py
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
arrow==1.3.0
    # via -r requirements/requirements.in
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via pycares
charset-normalizer==3.3.2
    # via requests
coloredlogs==15.0.1
    # via -r requirements/requirements.in
discord-py==2.3.2
    # via
    #   -r requirements/requirements.in
    #   pydis-core
filelock==3.13.4
    # via tldextract
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
humanfriendly==10.0
    # via coloredlogs
idna==3.7
    # via
    #   requests
    #   tldextract
    #   yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pycares==4.4.0
    # via aiodns
pycparser==2.22
    # via cffi
pydantic==2.7.0
    # via
    #   pydantic-settings
    #   pydis-core
pydantic-core==2.18.1
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements/requirements.in
pydis-core==11.1.0
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via arrow
python-dotenv==1.0.1
    # via pydantic-settings
rapidfuzz==3.8.1
    # via -r requirements/requirements.in
regex==2023.12.25
    # via -r requirements/requirements.in
requests==2.31.0
    # via
    #   requests-file
    #   tldextract
requests-file==2.0.0
    # via tldextract
sentry-sdk==1.45.0
    # via -r requirements/requirements.in
six==1.16.0
    # via python-dateutil
statsd==4.0.1
    # via pydis-core
tldextract==5.1.2
    # via -r requirements/requirements.in
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via
    #   requests
    #   sentry-sdk
yarl==1.9.4
    # via aiohttp
